Issue #294: Use upstream hostname for request
WIP WIP WIP WIP WIP WIP WIP WIP WIP WIP WIP WIP
-----------------------------------------------

This patch is work in progress and demonstrates the
desired behavior and adds a test to verify it. The
feature needs to be made configurable per route.

Fabio does not modify the Host header when forwarding
the request to the upstream server. This patch enables
a mode where for a specific route fabio will use the
hostname of the upstream server as the host header.

-----------------------------------------------
WIP WIP WIP WIP WIP WIP WIP WIP WIP WIP WIP WIP

Fixes #294